+++ server/database/migrations/2021_07_17_195642_create_sale_payments_table.php
@@ -18,8 +18,8 @@ class CreateSalePaymentsTable extends Migration
             $table->id();
             $table->string('reference')->default("INV/SL_1110");
 
-            $table->bigInteger('amount');
-            $table->bigInteger('payment_method')->default(Constants::PAYMENT_CASH);
+            $table->float('amount', 10, 0);
+            $table->integer('payment_method')->default(Constants::PAYMENT_CASH);
             $table->text('note')->nullable();
 
             $table->unsignedBigInteger('sale_id');
