+++ server/database/migrations/2021_07_18_065548_create_expenses_table.php
@@ -15,7 +15,7 @@ class CreateExpensesTable extends Migration
     {
         Schema::create('expenses', function (Blueprint $table) {
             $table->id();
-            $table->integer('amount');
+            $table->float('amount', 10, 0);
             $table->string('details');
             $table->date('date')->default(date('Y-m-d'));
             $table->string('reference')->default("EXP_1110");
